Реализовано модерирование комментариев

+++ resources/views/article/one_article.blade.php
@@ -47,18 +47,20 @@
     @endif
     <div class="comments-container">
         @foreach ($comments as $comment)
-            <div class="comment">
-                <p class="comment__author">{{ $comment->getAuthorName() }}</p>
-                <p class="comment__title">{{ $comment->title }}</p>
-                <p class="comment__text">{{ $comment->text }}</p>
-                <p class="comment__date">{{ $comment->created_at }}</p>
-                @can('comment', $comment)
-                    <div class="button-box">
-                        <a href="/comment/edit/{{ $comment->id }}" class="button button_blue">Редактировать</a>
-                        <a href="/comment/delete/{{ $comment->id }}" class="button button_red">Удалить</a>
-                    </div>
-                @endcan
-            </div>
+            @if ($comment->status)
+                <div class="comment">
+                    <p class="comment__title">{{ $comment->title }}</p>
+                    <p class="comment__text">{{ $comment->text }}</p>
+                    <p class="comment__date">{{ $comment->created_at }}</p>
+                    <p class="comment__author">{{ $comment->getAuthorName() }}</p>
+                    @can('comment', $comment)
+                        <div class="button-box">
+                            <a href="/comment/edit/{{ $comment->id }}" class="button button_blue">Редактировать</a>
+                            <a href="/comment/delete/{{ $comment->id }}" class="button button_red">Удалить</a>
+                        </div>
+                    @endcan
+                </div>
+            @endif
         @endforeach
     </div>
 </section>
